requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.3
requests==2.31.0
dnspython==2.4.2
email-validator==2.1.0
lxml==4.9.3
flask==3.0.0
flask-httpauth==4.8.0
gunicorn==21.2.0


